# V2/app/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tasker API",
    "description": "Manage projects, tasks, sections, and milestones in Tasker.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://smarttasksxdd.netlify.app/api/rest"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/tasks/{taskId}": {
      "patch": {
        "operationId": "updateTask",
        "summary": "Update a task",
        "parameters": [
          { "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": { "type": "string" },
                  "detail": { "type": "string" },
                  "priority": { "type": "string", "enum": ["rush", "high", "medium", "low"] },
                  "due_date": { "type": "string", "format": "date" },
                  "status": { "type": "string", "enum": ["pending", "in_progress", "done"] },
                  "section_id": { "type": "string" },
                  "pinned": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated task" }
        }
      },
      "delete": {
        "operationId": "deleteTask",
        "summary": "Delete a task",
        "parameters": [
          { "name": "taskId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted" }
        }
      }
    },
...
  }
}